Include cast to int in imshow_rand

+++ gala/viz.py
@@ -55,7 +55,7 @@ def imshow_rand(im, labrandom=True):
     fig : plt.Figure
         The image shown.
     """
-    rand_colors = np.random.rand(np.ceil(im.max()), 3)
+    rand_colors = np.random.rand(int(np.ceil(im.max())), 3)
     if labrandom:
         rand_colors[:, 0] = rand_colors[:, 0] * 60 + 20
         rand_colors[:, 1] = rand_colors[:, 1] * 185 - 85
